pyproject.toml
[project]
name = "superhero-max"
version = "0.1.0"
description = "AI-powered Superhero Avatar Generator - Transform photos into personalized superhero avatars using Streamlit and Replicate"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.28.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "replicate>=0.15.0",
    "pandas>=2.0.0",
    "boto3>=1.28.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "databricks-sdk>=0.59.0",
    "opencv-python>=4.12.0.88",
    "fal-client>=0.7.0",
    "email-validator>=2.0.0",
    "google-cloud-storage>=2.10.0",
    "qrcode[pil]>=7.4.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
